fix freebsd build
