+++ pkg/metadata/v5/v5_other.go
@@ -10,18 +10,16 @@ package v5
 import (
 	"github.com/DataDog/datadog-agent/pkg/metadata/common"
 	"github.com/DataDog/datadog-agent/pkg/metadata/host"
-	"github.com/DataDog/datadog-agent/pkg/metadata/resources"
 )
 
-// GetPayload returns the complete metadata payload as seen in Agent v5
+// GetPayload returns the complete metadata payload as seen in Agent v5.
+// Note: gohai can't be used on the platforms this module builds for
 func GetPayload(hostname string) *Payload {
 	cp := common.GetPayload(hostname)
 	hp := host.GetPayload(hostname)
-	rp := resources.GetPayload(hostname)
 
 	return &Payload{
-		CommonPayload:    CommonPayload{*cp},
-		HostPayload:      HostPayload{*hp},
-		ResourcesPayload: ResourcesPayload{*rp},
+		CommonPayload: CommonPayload{*cp},
+		HostPayload:   HostPayload{*hp},
 	}
 }
