+++ pkg/status/platform_no_gohai.go
@@ -1,0 +1,12 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2018 Datadog, Inc.
+
+// +build freebsd netbsd openbsd solaris dragonfly
+
+package status
+
+func getPlatformPayload() (result interface{}, err error) {
+	return nil, nil
+}
